move the firebaseConfig to context file
