+++ package.json
@@ -21,7 +21,7 @@
     "@types/react-redux": "^7.1.33",
     "crypto-browserify": "^3.12.0",
     "dayjs": "^1.11.10",
-    "firebase": "^10.9.0",
+    "firebase": "^10.10.0",
     "moment": "^2.30.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
